+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   structures so that the complete result of fitting a box is returned
 * [HexaPDF::Layout::Frame] to use a better algorithm for trimming the shape
 * [HexaPDF::Layout::Frame::new] to allow setting the initial shape
+* **Breaking change**: Removed contour line from [HexaPDF::Layout::Frame]
 
 ### Fixed